Component
This update is rare based on a function that is very common illustrating a mechanism that is consider unethical to many but it holds a key to open a chest full of ideas. This program is currently under maintaince until the next update v0.5 where the program will be able to display both grid and values asked without crashing.